% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoQR.R
\name{autoplot.CoQR}
\alias{autoplot.CoQR}
\title{Autoplot a CoQR object}
\usage{
\method{autoplot}{CoQR}(obj, facet_names = c("X / VaR", "Y / CoVaR"))
}
\arguments{
\item{obj}{CoQR object}

\item{facet_names}{Titles of the facets}
}
\value{
A ggplot object
}
\description{
Autoplot a CoQR object
}
