% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoQRroll.R
\name{autoplot.CoQRroll}
\alias{autoplot.CoQRroll}
\title{CoQRroll autoplot method}
\usage{
\method{autoplot}{CoQRroll}(obj, facet_names = c("X / VaR", "Y / CoVaR"))
}
\arguments{
\item{facet_names}{}
}
\value{

}
\description{
CoQRroll autoplot method
}
