% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoQR-package.R
\docType{package}
\name{CoQR-package}
\alias{CoQR}
\alias{CoQR-package}
\title{CoQR: Joint Dynamic Models for the VaR and CoVaR}
\description{
The package estimates joint dynamic autoregressive or cross-sectional models of the VaR (Value-at-Risk) (i.e., the quantile) and the CoVaR of a bivariate vector of financial returns.
}
\author{
\strong{Maintainer}: Timo Dimitriadis \email{timo.dimi@gmail.com} (\href{https://orcid.org/0000-0002-8322-360X}{ORCID})

Authors:
\itemize{
  \item Yannick Hoga \email{Yannick.Hoga@vwl.uni-due.de}
}

}
\keyword{internal}
