% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoQR.R
\name{CoQR.default}
\alias{CoQR.default}
\title{Title}
\usage{
\method{CoQR}{default}(
  data = NULL,
  x = NULL,
  y = NULL,
  z = NULL,
  model = "CoCAViaR_SAV_diag",
  SRM = "CoVaR",
  beta = 0.95,
  alpha = 0.95,
  theta0 = NULL,
  optim_replications = c(1, 3)
)
}
\arguments{
\item{optim_replications}{}
}
\value{

}
\description{
Title
}
