% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoQR.R
\name{CoQR.fit}
\alias{CoQR.fit}
\title{Title}
\usage{
\method{CoQR}{fit}(data, x, y, z, model, SRM, beta, alpha, theta0, optim_replications)
}
\arguments{
\item{optim_replications}{}
}
\value{

}
\description{
Title
}
