% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoQRroll.R
\name{CoQRroll}
\alias{CoQRroll}
\title{Title}
\usage{
CoQRroll(
  data,
  model = "CoCAViaR_SAV_diag",
  length_IS = 1000,
  refit_freq = 100,
  SRM = "CoVaR",
  beta = 0.95,
  alpha = 0.95,
  theta0 = NULL,
  optim_replications = c(1, 3)
)
}
\arguments{
\item{optim_replications}{}
}
\value{

}
\description{
Title
}
