% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoQR.R
\name{cov.CoQR}
\alias{cov.CoQR}
\title{CoQR p arameter covariance estimation}
\usage{
cov.CoQR(CoQR_object, method, ...)
}
\arguments{
\item{CoQR_object}{A CoQR object}

\item{method}{The method to compute the standard errors. Either "asymptotic" or "boot"}

\item{...}{Further input parameters}
}
\value{
covariance matrix
}
\description{
CoQR p arameter covariance estimation
}
