% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoQR.R
\name{forecast.CoQR}
\alias{forecast.CoQR}
\title{CoQR forecast method}
\usage{
\method{forecast}{CoQR}(obj, newdata = NULL)
}
\arguments{
\item{obj}{CoQR object}

\item{newdata}{new data set to use for forecasting (if not re-estimated in this time step)}
}
\value{
Vector of forecasts
}
\description{
CoQR forecast method
}
