% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoQR.R
\name{print.CoQR}
\alias{print.CoQR}
\title{print method for the CoQR class}
\usage{
\method{print}{CoQR}(obj, digits = 4)
}
\arguments{
\item{obj}{CoQR object}

\item{digits}{printed digits after the comma}
}
\description{
print method for the CoQR class
}
