% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoQR.R
\name{print.summary.CoQR}
\alias{print.summary.CoQR}
\title{print method for the sum.CoQR object}
\usage{
\method{print}{summary.CoQR}(sum.CoQR_object)
}
\arguments{
\item{sum.CoQR_object}{A sum.CoQR object}
}
\value{
Nothing
}
\description{
print method for the sum.CoQR object
}
