% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemicRiskFCeval.R
\name{print.SystemicRiskFCeval}
\alias{print.SystemicRiskFCeval}
\title{Title}
\usage{
\method{print}{SystemicRiskFCeval}(obj, digits = 4)
}
\arguments{
\item{digits}{}
}
\value{

}
\description{
Title
}
