% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoQR.R
\name{summary.CoQR}
\alias{summary.CoQR}
\title{CoQR summary method}
\usage{
\method{summary}{CoQR}(CoQR_object, method = "asymptotic", ...)
}
\arguments{
\item{CoQR_object}{A CoQR object}

\item{method}{The method to compute the standard errors. Either "asymptotic" or "boot"}

\item{...}{Further input parameters}
}
\value{
A summary.CoQR object
}
\description{
CoQR summary method
}
