% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemicRiskFCeval.R
\name{SystemicRiskFCeval}
\alias{SystemicRiskFCeval}
\title{Title}
\usage{
SystemicRiskFCeval(
  x = NULL,
  y = NULL,
  VaR1 = NULL,
  VaR2 = NULL,
  CoVaR1 = NULL,
  CoVaR2 = NULL,
  MES1 = NULL,
  MES2 = NULL,
  data = NULL,
  SRM = NULL,
  beta = 0.95,
  alpha = 0.95,
  sided = "onehalf",
  cov_method = "HAC",
  sig_level = 0.05
)
}
\arguments{
\item{sig_level}{}
}
\value{

}
\description{
Title
}
