% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoQR.R
\name{autoplot.CoQR}
\alias{autoplot.CoQR}
\title{Title}
\usage{
\method{autoplot}{CoQR}(obj, facet_names = c("X / VaR", "Y / CoVaR"))
}
\arguments{
\item{facet_names}{}
}
\value{

}
\description{
Title
}
\examples{

}
