% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoQRroll.R
\name{CoQRroll}
\alias{CoQRroll}
\title{Rolling Forecasts for Dynamic (VaR, CoVaR) Models}
\usage{
CoQRroll(
  data,
  model = "CoCAViaR_SAV_diag",
  length_IS = 1000,
  refit_freq = 100,
  SRM = "CoVaR",
  beta = 0.95,
  alpha = 0.95,
  theta0 = NULL,
  optim_replications = c(1, 3)
)
}
\arguments{
\item{optim_replications}{}
}
\value{

}
\description{
Rolling Forecasts for Dynamic (VaR, CoVaR) Models
}
\examples{
Simulate bivariate GARCH data
library(rmgarch)
data(dji30retw)

# Estimate the "CoCAViaR_SAV_diag" model on the negative percentage log-returns
obj_roll <- CoQRroll(data=data.frame(x=-100*as.numeric(dji30retw$JPM),
y=-100*rowMeans(dji30retw)),
model="CoCAViaR_SAV_diag",
length_IS=1000, refit_freq=100)

# Covariance estimation and display the parameter estimates with standard errors
summary(obj_roll)

# Plot the forecasts
plot (obj_roll)

}
\references{
\href{https://arxiv.org/abs/.....}{A Dynamic Co-Quantile Regression}
}
